# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Regulatory Tracking Dashboard API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type RegulationJurisdiction =
  (typeof RegulationJurisdiction)[keyof typeof RegulationJurisdiction];

export const RegulationJurisdiction = {
  federal: "federal",
  state: "state",
  international: "international",
} as const;

export type RegulationStatus =
  (typeof RegulationStatus)[keyof typeof RegulationStatus];

export const RegulationStatus = {
  proposed: "proposed",
  enacted: "enacted",
  effective: "effective",
} as const;

export interface Regulation {
  id: number;
  lawName: string;
  jurisdiction: RegulationJurisdiction;
  status: RegulationStatus;
  summary: string;
  relevantPolicies: string[];
  nextAction: string;
  effectiveDate?: string | null;
  deadlineDate?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type CreateRegulationInputJurisdiction =
  (typeof CreateRegulationInputJurisdiction)[keyof typeof CreateRegulationInputJurisdiction];

export const CreateRegulationInputJurisdiction = {
  federal: "federal",
  state: "state",
  international: "international",
} as const;

export type CreateRegulationInputStatus =
  (typeof CreateRegulationInputStatus)[keyof typeof CreateRegulationInputStatus];

export const CreateRegulationInputStatus = {
  proposed: "proposed",
  enacted: "enacted",
  effective: "effective",
} as const;

export interface CreateRegulationInput {
  lawName: string;
  jurisdiction: CreateRegulationInputJurisdiction;
  status: CreateRegulationInputStatus;
  summary: string;
  relevantPolicies: string[];
  nextAction: string;
  effectiveDate?: string | null;
  deadlineDate?: string | null;
}

export type RegulationStatsByJurisdictionItem = {
  jurisdiction: string;
  count: number;
};

export type RegulationStatsByStatusItem = {
  status: string;
  count: number;
};

export interface RegulationStats {
  totalCount: number;
  byJurisdiction: RegulationStatsByJurisdictionItem[];
  byStatus: RegulationStatsByStatusItem[];
  upcomingDeadlines: Regulation[];
}

export interface FrameworkControl {
  id: string;
  name: string;
  description: string;
}

export interface FrameworkCategory {
  id: string;
  name: string;
  description: string;
  controls: FrameworkControl[];
}

export interface Framework {
  id: string;
  name: string;
  description: string;
  categories: FrameworkCategory[];
}

export type PolicyStatus = (typeof PolicyStatus)[keyof typeof PolicyStatus];

export const PolicyStatus = {
  draft: "draft",
  active: "active",
  under_review: "under_review",
  retired: "retired",
} as const;

export interface Policy {
  id: number;
  name: string;
  description: string;
  coveredControls: string[];
  status: PolicyStatus;
  lastReviewed?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type CreatePolicyInputStatus =
  (typeof CreatePolicyInputStatus)[keyof typeof CreatePolicyInputStatus];

export const CreatePolicyInputStatus = {
  draft: "draft",
  active: "active",
  under_review: "under_review",
  retired: "retired",
} as const;

export interface CreatePolicyInput {
  name: string;
  description: string;
  coveredControls: string[];
  status: CreatePolicyInputStatus;
  lastReviewed?: string | null;
}

export interface GapAnalysisInput {
  frameworkId: string;
}

export interface GapItem {
  controlId: string;
  controlName: string;
  categoryName: string;
  description: string;
  coveredByPolicies: string[];
}

export interface GapAnalysisResult {
  frameworkId: string;
  frameworkName: string;
  totalControls: number;
  coveredControls: number;
  gapControls: number;
  coveragePercent: number;
  gaps: GapItem[];
  covered: GapItem[];
}

export type ListRegulationsParams = {
  jurisdiction?: string;
  status?: string;
  search?: string;
};