# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Regulatory Tracking Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /gap-analysis/policies:
    get:
      operationId: listPolicies
      tags: [gap-analysis]
      summary: List organization policies
      responses:
        "200":
          description: List of policies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Policy"
    post:
      operationId: createPolicy
      tags: [gap-analysis]
      summary: Create an organization policy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePolicyInput"
      responses:
        "201":
          description: Created policy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Policy"
  /gap-analysis/policies/{id}:
    put:
      operationId: updatePolicy
      tags: [gap-analysis]
      summary: Update a policy
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePolicyInput"
      responses:
        "200":
          description: Updated policy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Policy"
    delete:
      operationId: deletePolicy
      tags: [gap-analysis]
      summary: Delete a policy
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /gap-analysis/analyze:
    post:
      operationId: analyzeGaps
      tags: [gap-analysis]
      summary: Analyze policy gaps against a framework
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GapAnalysisInput"
      responses:
        "200":
          description: Gap analysis results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GapAnalysisResult"
components:
  schemas:
...
    Policy:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        coveredControls:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [draft, active, under_review, retired]
        lastReviewed:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - description
        - coveredControls
        - status
        - createdAt
        - updatedAt
    CreatePolicyInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        coveredControls:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [draft, active, under_review, retired]
        lastReviewed:
          type: string
          nullable: true
      required:
        - name
        - description
        - coveredControls
        - status
    GapAnalysisInput:
      type: object
      properties:
        frameworkId:
          type: string
      required:
        - frameworkId
    GapAnalysisResult:
      type: object
      properties:
        frameworkId:
          type: string
        frameworkName:
          type: string
        totalControls:
          type: integer
        coveredControls:
          type: integer
        gapControls:
          type: integer
        coveragePercent:
          type: number
        gaps:
          type: array
          items:
            $ref: "#/components/schemas/GapItem"
        covered:
          type: array
          items:
            $ref: "#/components/schemas/GapItem"
      required:
        - frameworkId
        - frameworkName
        - totalControls
        - coveredControls
        - gapControls
        - coveragePercent
        - gaps
        - covered
    GapItem:
      type: object
      properties:
        controlId:
          type: string
        controlName:
          type: string
        categoryName:
          type: string
        description:
          type: string
        coveredByPolicies:
          type: array
          items:
            type: string
      required:
        - controlId
        - controlName
        - categoryName
        - description
        - coveredByPolicies