# lib/api-zod/src/generated/types/createPolicyInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Regulatory Tracking Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { CreatePolicyInputStatus } from "./createPolicyInputStatus";

export interface CreatePolicyInput {
  name: string;
  description: string;
  coveredControls: string[];
  status: CreatePolicyInputStatus;
  lastReviewed?: string | null;
}