# lib/api-zod/src/generated/types/gapAnalysisInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Regulatory Tracking Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface GapAnalysisInput {
  frameworkId: string;
}