# lib/api-zod/src/generated/types/gapAnalysisResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Regulatory Tracking Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { GapItem } from "./gapItem";

export interface GapAnalysisResult {
  frameworkId: string;
  frameworkName: string;
  totalControls: number;
  coveredControls: number;
  gapControls: number;
  coveragePercent: number;
  gaps: GapItem[];
  covered: GapItem[];
}