# lib/api-zod/src/generated/types/gapItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Regulatory Tracking Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface GapItem {
  controlId: string;
  controlName: string;
  categoryName: string;
  description: string;
  coveredByPolicies: string[];
}