# lib/api-zod/src/generated/types/policy.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Regulatory Tracking Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { PolicyStatus } from "./policyStatus";

export interface Policy {
  id: number;
  name: string;
  description: string;
  coveredControls: string[];
  status: PolicyStatus;
  lastReviewed?: string | null;
  createdAt: string;
  updatedAt: string;
}